(core) Create core package; move cli_config/cli_matching/cli_image
