+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog is initialized from git commit history after `v1.0.0` and can be 
 ## [Unreleased]
 
 ### Changed
+- **`core`-paket infört; `cli_config`/`cli_matching`/`cli_image` flyttade dit (move-only).** Audit-steg 1 av att lyfta ut delad logik ur CLI-modullagret: modulerna flyttades via `git mv` till ett nytt `backend/core/`-paket (`core/config.py`, `core/matching.py`, `core/image.py`) utan någon logik- eller funktionsnamnsändring. De gamla sökvägarna finns kvar som tunna bakåtkompatibilitetsskal (`backend/cli_config.py` m.fl. → `from core.<modul> import *`) så befintliga importörer (`hitta_ansikten.py`, `api/services/culling_service.py`, arkiverade skript) fortsätter fungera oförändrat. Paketering uppdaterad: `packages = ["core"]` i `backend/pyproject.toml` och `core.*` tillagt i PyInstaller-specens `hiddenimports`. Culling-retentiontestet pekar nu på `core.config` i stället för skalet (monkeypatch måste träffa den riktiga modulen där `load_config` läser `BASE_DIR`/`CONFIG_PATH`).
 - **Orphan CLI-skript arkiverade till `backend/scripts/archive/`.** Åtta fristående legacy-/engångsverktyg utan runtime-importörer (verifierat via grep i audit 2026-07) flyttades ut ur `backend/`-roten: `analysera_ansikten.py`, `hantera_ansikten.py`, `forfina_ansikten.py`, `ratta_ansikten.py`, `rensa_dlib.py`, `migrera_processed.py`, `inspect_encodings.py` och `update_encodings_with_filehash.py`. Inget raderades — de förblir körbara. Varje flyttat skript som importerar backend-moduler (t.ex. `faceid_db`) fick en liten `sys.path`-bootstrap i toppen som lägger till `backend/` (tre nivåer upp), så de kan köras direkt: `python scripts/archive/<verktyg>.py` från `backend/`. En `backend/scripts/archive/README.md` beskriver vad de är, varför de arkiverades och hur de körs. Följdändring: dokumentationens sökvägar till `rensa_dlib.py` m.fl. pekar nu på `scripts/archive/…`. `filer2mappar.py`, `rakna_spelare.py`, `rename_nef.py` och `hitta_ansikten.py` är fortsatt aktiva och lämnades i `backend/`-roten.
 
 ### Removed
