+++ backend/core/__init__.py
@@ -1,0 +1,5 @@
+"""Core package: shared logic extracted from the CLI-module layer.
+
+Houses configuration, face-matching and image utilities used by both the
+legacy CLI (``hitta_ansikten.py``) and the FastAPI backend (``api/``).
+"""
